+++ src/main/java/ru/javawebinar/topjava/repository/MealRepository.java
@@ -20,8 +20,4 @@ public interface MealRepository {
 
     // ORDERED dateTime desc
     List<Meal> getBetweenHalfOpen(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId);
-
-    default Meal getWithUser(int id, int userId) {
-        throw new UnsupportedOperationException();
-    }
 }
