+++ src/main/java/ru/javawebinar/topjava/repository/UserRepository.java
@@ -18,8 +18,4 @@ public interface UserRepository {
     User getByEmail(String email);
 
     List<User> getAll();
-
-    default User getWithMeals(int id) {
-        throw new UnsupportedOperationException();
-    }
 }
