+++ src/main/java/ru/javawebinar/topjava/repository/datajpa/CrudUserRepository.java
@@ -16,7 +16,4 @@ public interface CrudUserRepository extends JpaRepository<User, Integer> {
     int delete(@Param("id") int id);
 
     User getByEmail(String email);
-
-    @Query("SELECT u FROM User u LEFT JOIN FETCH u.meals WHERE u.id=:id")
-    User getWithMeals(@Param("id") int id);
 }
