+++ src/main/java/ru/javawebinar/topjava/repository/datajpa/DataJpaUserRepository.java
@@ -41,9 +41,4 @@ public class DataJpaUserRepository implements UserRepository {
     public List<User> getAll() {
         return crudRepository.findAll(SORT_NAME_EMAIL);
     }
-
-    @Override
-    public User getWithMeals(int id) {
-        return crudRepository.getWithMeals(id);
-    }
 }
