--- src/main/java/ru/javawebinar/topjava/repository/jdbc/HsqlJdbcMealRepository.java
@@ -1,23 +1,0 @@
-package ru.javawebinar.topjava.repository.jdbc;
-
-import org.springframework.context.annotation.Profile;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Repository;
-import ru.javawebinar.topjava.Profiles;
-
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-
-@Repository
-@Profile(Profiles.HSQL_DB)
-public class HsqlJdbcMealRepository extends JdbcMealRepository<Timestamp> {
-    public HsqlJdbcMealRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
-        super(jdbcTemplate, namedParameterJdbcTemplate);
-    }
-
-    @Override
-    protected Timestamp getDateTime(LocalDateTime dateTime) {
-        return Timestamp.valueOf(dateTime);
-    }
-}
