--- src/main/java/ru/javawebinar/topjava/repository/jdbc/PostgresJdbcMealRepository.java
@@ -1,22 +1,0 @@
-package ru.javawebinar.topjava.repository.jdbc;
-
-import org.springframework.context.annotation.Profile;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Repository;
-import ru.javawebinar.topjava.Profiles;
-
-import java.time.LocalDateTime;
-
-@Repository
-@Profile(Profiles.POSTGRES_DB)
-public class PostgresJdbcMealRepository extends JdbcMealRepository<LocalDateTime> {
-    public PostgresJdbcMealRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
-        super(jdbcTemplate, namedParameterJdbcTemplate);
-    }
-
-    @Override
-    protected LocalDateTime getDateTime(LocalDateTime dateTime) {
-        return dateTime;
-    }
-}
