+++ src/main/java/ru/javawebinar/topjava/service/MealService.java
@@ -47,8 +47,4 @@ public class MealService {
         Assert.notNull(meal, "meal must not be null");
         return repository.save(meal, userId);
     }
-
-    public Meal getWithUser(int id, int userId) {
-        return checkNotFoundWithId(repository.getWithUser(id, userId), id);
-    }
 }
