+++ src/main/java/ru/javawebinar/topjava/service/UserService.java
@@ -51,8 +51,4 @@ public class UserService {
         Assert.notNull(user, "user must not be null");
         checkNotFoundWithId(repository.save(user), user.id());
     }
-
-    public User getWithMeals(int id) {
-        return checkNotFoundWithId(repository.getWithMeals(id), id);
-    }
 }
