--- src/test/java/ru/javawebinar/topjava/service/datajpa/DataJpaMealServiceTest.java
@@ -1,20 +1,0 @@
-package ru.javawebinar.topjava.service.datajpa;
-
-import org.junit.Test;
-import org.springframework.test.context.ActiveProfiles;
-import ru.javawebinar.topjava.Profiles;
-import ru.javawebinar.topjava.model.Meal;
-import ru.javawebinar.topjava.service.MealServiceTest;
-
-import static ru.javawebinar.topjava.MealTestData.*;
-import static ru.javawebinar.topjava.UserTestData.*;
-
-@ActiveProfiles(profiles = Profiles.DATAJPA)
-public class DataJpaMealServiceTest extends MealServiceTest {
-    @Test
-    public void getWithUser() {
-        Meal actual = service.getWithUser(ADMIN_MEAL_ID, ADMIN_ID);
-        MEAL_MATCHER.assertMatch(actual, adminMeal1);
-        USER_MATCHER.assertMatch(actual.getUser(), admin);
-    }
-}
