--- src/test/java/ru/javawebinar/topjava/service/datajpa/DataJpaUserServiceTest.java
@@ -1,20 +1,0 @@
-package ru.javawebinar.topjava.service.datajpa;
-
-import org.junit.Test;
-import org.springframework.test.context.ActiveProfiles;
-import ru.javawebinar.topjava.Profiles;
-import ru.javawebinar.topjava.model.User;
-import ru.javawebinar.topjava.service.UserServiceTest;
-
-import static ru.javawebinar.topjava.MealTestData.*;
-import static ru.javawebinar.topjava.UserTestData.*;
-
-@ActiveProfiles(profiles = Profiles.DATAJPA)
-public class DataJpaUserServiceTest extends UserServiceTest {
-    @Test
-    public void getWithMeals() {
-        User actual = service.getWithMeals(ADMIN_ID);
-        USER_MATCHER.assertMatch(actual, admin);
-        MEAL_MATCHER.assertMatch(actual.getMeals(), adminMeal2, adminMeal1);
-    }
-}
