--- src/test/java/ru/javawebinar/topjava/service/jdbc/JdbcMealServiceTest.java
@@ -1,9 +1,0 @@
-package ru.javawebinar.topjava.service.jdbc;
-
-import org.springframework.test.context.ActiveProfiles;
-import ru.javawebinar.topjava.Profiles;
-import ru.javawebinar.topjava.service.MealServiceTest;
-
-@ActiveProfiles(profiles = Profiles.JDBC)
-public class JdbcMealServiceTest extends MealServiceTest {
-}
