--- src/test/java/ru/javawebinar/topjava/service/jdbc/JdbcUserServiceTest.java
@@ -1,9 +1,0 @@
-package ru.javawebinar.topjava.service.jdbc;
-
-import org.springframework.test.context.ActiveProfiles;
-import ru.javawebinar.topjava.Profiles;
-import ru.javawebinar.topjava.service.UserServiceTest;
-
-@ActiveProfiles(profiles = Profiles.JDBC)
-public class JdbcUserServiceTest extends UserServiceTest {
-}
