--- src/test/java/ru/javawebinar/topjava/service/jpa/JpaMealServiceTest.java
@@ -1,9 +1,0 @@
-package ru.javawebinar.topjava.service.jpa;
-
-import org.springframework.test.context.ActiveProfiles;
-import ru.javawebinar.topjava.Profiles;
-import ru.javawebinar.topjava.service.MealServiceTest;
-
-@ActiveProfiles(profiles = Profiles.JPA)
-public class JpaMealServiceTest extends MealServiceTest {
-}
