--- src/test/java/ru/javawebinar/topjava/service/jpa/JpaUserServiceTest.java
@@ -1,9 +1,0 @@
-package ru.javawebinar.topjava.service.jpa;
-
-import org.springframework.test.context.ActiveProfiles;
-import ru.javawebinar.topjava.Profiles;
-import ru.javawebinar.topjava.service.UserServiceTest;
-
-@ActiveProfiles(profiles = Profiles.JPA)
-public class JpaUserServiceTest extends UserServiceTest {
-}
