+++ src/test/resources/logback-test.xml
@@ -24,7 +24,6 @@
     <!--    <logger name="org.postgresql" level="debug"/>-->
     <!--    <logger name="org.springframework.jdbc" level="info"/>-->
     <logger name="ru.javawebinar.topjava" level="debug"/>
-    <logger name="org.springframework.orm.jpa" level="debug"/>
 
     <root level="INFO">
         <appender-ref ref="console"/>
